Separate lexer and parser from cfront
